refactor(serve): rename `franklin webui` command → `franklin serve`
Drops the webui naming (the local agent server backs the desktop app, not just
a browser UI). src/webui → src/serve, webuiCommand → serveCommand,
startWebuiServer → startServer.
