+++ src/index.ts
@@ -116,14 +116,14 @@ program
   });
 
 program
-  .command('webui')
-  .description('Run the local WebUI/desktop agent server (WebSocket on localhost:3737)')
+  .command('serve')
+  .description('Run the local agent server for the desktop app / browser UI (WebSocket on localhost:3737)')
   .option('-p, --port <port>', 'Agent server port', '3737')
   .option('--work-dir <dir>', 'Working directory for tools (default: cwd)')
   .option('--debug', 'Verbose logging')
   .action(async (options: { port?: string; workDir?: string; debug?: boolean }) => {
-    const { webuiCommand } = await import('./commands/webui.js');
-    await webuiCommand(options);
+    const { serveCommand } = await import('./commands/serve.js');
+    await serveCommand(options);
   });
 
 program
